Add javadoc Remove unused import
Add javadoc for all class in Commands package.
Remove unused import in both Commands package and Activities package.
