+++ src/main/java/Duke/Commands/AddEvent.java
@@ -5,17 +5,37 @@ import Duke.Activities.ActivityList;
 import Duke.Converstion.Dialog;
 import Duke.Exception.CommandException;
 
-import javax.print.attribute.standard.MediaSize;
 import java.time.LocalDate;
-import java.util.Arrays;
-import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class AddEvent extends Command implements AddActivity{
+/**
+ * Represents a command to add an Event activity to an {@link ActivityList}.
+ * Extends the {@link Command} class and implements the {@link AddActivity} interface.
+ */
+public class AddEvent extends Command implements AddActivity {
+
+    /**
+     * The name of the Event activity.
+     */
     private final String NAME;
+
+    /**
+     * The start date of the Event activity.
+     */
     private final LocalDate STARTDATE;
+
+    /**
+     * The end date of the Event activity.
+     */
     private final LocalDate ENDDATE;
+
+    /**
+     * Constructor to initialize the AddEvent command by parsing the input.
+     *
+     * @param input The input string containing the name, start date, and end date of the Event activity.
+     * @throws CommandException If the input format is incorrect or if the start date is after the end date.
+     */
     public AddEvent(String input) throws CommandException {
         super(input);
         String[] phrased = phrases(input);
@@ -23,6 +43,14 @@ public class AddEvent extends Command implements AddActivity{
         STARTDATE = DateFormat.format(phrased[1]);
         ENDDATE = DateFormat.format(phrased[2]);
     }
+
+    /**
+     * Parses the input string to extract the name, start date, and end date of the Event activity.
+     *
+     * @param input The input string containing the name, start date, and end date.
+     * @return An array containing the name at index 0, start date at index 1, and end date at index 2.
+     * @throws CommandException If the input format is incorrect.
+     */
     public String[] phrases(String input) throws CommandException {
         String patternRegex = "^[^/]+ /from \\d{4}-\\d{1,2}-\\d{1,2} /to \\d{4}-\\d{1,2}-\\d{1,2}$";
         Pattern pattern = Pattern.compile(patternRegex);
@@ -36,15 +64,27 @@ public class AddEvent extends Command implements AddActivity{
             String toDate = dateStr[1];
             return new String[]{name, fromDate, toDate};
         } else {
-            throw new CommandException("The input has to be in prompt: \nname /from yyyy-mm-dd /to yyyy-mm-dd");
+            throw new CommandException("The input has to be in the format: \nname /from yyyy-mm-dd /to yyyy-mm-dd");
         }
     }
 
+    /**
+     * Adds the Event activity to the specified {@link ActivityList}.
+     *
+     * @param list The activity list to which the Event activity should be added.
+     * @return The added Event activity.
+     */
     @Override
     public Activity addToList(ActivityList list) {
         return list.add(NAME, STARTDATE, ENDDATE);
     }
 
+    /**
+     * Executes the command by adding the Event activity to the specified {@link ActivityList}.
+     *
+     * @param list The activity list to which the Event activity should be added.
+     * @throws CommandException If an error occurs during execution, such as the start date being after the end date.
+     */
     @Override
     public void execute(ActivityList list) throws CommandException {
         if (DateFormat.compareDate(STARTDATE, ENDDATE)) {
@@ -54,6 +94,11 @@ public class AddEvent extends Command implements AddActivity{
         }
     }
 
+    /**
+     * Returns a string representation of the command's execution output.
+     *
+     * @return A string indicating the successful addition of the Event activity.
+     */
     @Override
     public String toString() {
         String output = Dialog.printLine();
